soundwire: add initial definitions for sdw_master_device
Since we want an explicit support for the SoundWire Master device, add
the definitions, following the Greybus example of a 'Host Device'.

A parent (such as the Intel audio controller) would use
sdw_master_device_add() to create the device, passing a driver as a
parameter. The sdw_master_device_release() would be called when
put_device() is invoked by the parent. As requested by maintainers we
do not use shortcuts for the master device.

Note that since there is no standard for the Master host-facing
interface, there is no core which automagically loads the drivers. The
integration layers must manually call sdw_master_device_add(), in the
same way the platform_devices needed to be created so far.

The 'Master Device' driver exposes interfaces for
probe/startup/shutdown/remove and a placeholder for autonomous clock
stop support (when the bus control is handed over to a lower-power
solution, typically in D0ix modes).

Module namespace support is added in a separate patch.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit 4228e02411d8024b73999bf805e31d1b829ad8ed)
Signed-off-by: Hui Wang <hui.wang@canonical.com>