ASoC: SOF: add core id to sof_ipc_comp
Adds core id to sof_ipc_comp. The intention of this change
is to inform FW on which core that particular component
should run. Right now core id is only passed when pipeline
is created, which is not flexible enough and doesn't allow
for FW to handle this the right way.

Signed-off-by: Tomasz Lauda <tomasz.lauda@linux.intel.com>
(cherry picked from commit 0edd5e08bc2ffd1e503de73fc8fb55fc068bd6d9)
Signed-off-by: Hui Wang <hui.wang@canonical.com>