soundwire: cadence: remove automatic command retries
This is a good idea on paper, but it's not recommended at all when
operating in multi-master mode. It's also not recommended when doing
bank switches, since the retransmission would happen at the next SSP,
and the command protocol is stuck in the mean time.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit 44d14c80f0ee78086d488f7d512dc7dbdd580d88)
Signed-off-by: Hui Wang <hui.wang@canonical.com>