ASoC: Intel: broadwell: Fix oops during module removal
When removing the SOF module, the RT286 jack detect
handler will oops if jack detection is not disabled.
Disable the jack in the machine driver remove callback
to prevent this. This fix is only for SOF support and is
not needed for earlier versions.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
(cherry picked from commit 3611788300ce41daa6b0c88453ba0132898a21ac)
Signed-off-by: Hui Wang <hui.wang@canonical.com>